feat: reorganise prompts
Splits up the prompts module into separate modules.
