+++ packages/prompts/src/group.ts
@@ -1,0 +1,57 @@
+import { isCancel } from '@clack/core';
+
+type Prettify<T> = {
+	[P in keyof T]: T[P];
+} & {};
+
+export type PromptGroupAwaitedReturn<T> = {
+	[P in keyof T]: Exclude<Awaited<T[P]>, symbol>;
+};
+
+export interface PromptGroupOptions<T> {
+	/**
+	 * Control how the group can be canceled
+	 * if one of the prompts is canceled.
+	 */
+	onCancel?: (opts: {
+		results: Prettify<Partial<PromptGroupAwaitedReturn<T>>>;
+	}) => void;
+}
+
+export type PromptGroup<T> = {
+	[P in keyof T]: (opts: {
+		results: Prettify<Partial<PromptGroupAwaitedReturn<Omit<T, P>>>>;
+	}) => undefined | Promise<T[P] | undefined>;
+};
+
+/**
+ * Define a group of prompts to be displayed
+ * and return a results of objects within the group
+ */
+export const group = async <T>(
+	prompts: PromptGroup<T>,
+	opts?: PromptGroupOptions<T>
+): Promise<Prettify<PromptGroupAwaitedReturn<T>>> => {
+	const results = {} as any;
+	const promptNames = Object.keys(prompts);
+
+	for (const name of promptNames) {
+		const prompt = prompts[name as keyof T];
+		const result = await prompt({ results })?.catch((e) => {
+			throw e;
+		});
+
+		// Pass the results to the onCancel function
+		// so the user can decide what to do with the results
+		// TODO: Switch to callback within core to avoid isCancel Fn
+		if (typeof opts?.onCancel === 'function' && isCancel(result)) {
+			results[name] = 'canceled';
+			opts.onCancel({ results });
+			continue;
+		}
+
+		results[name] = result;
+	}
+
+	return results;
+};
