+++ packages/prompts/src/messages.ts
@@ -1,0 +1,18 @@
+import type { Writable } from 'node:stream';
+import color from 'picocolors';
+import { type CommonOptions, S_BAR, S_BAR_END, S_BAR_START } from './common.js';
+
+export const cancel = (message = '', opts?: CommonOptions) => {
+	const output: Writable = opts?.output ?? process.stdout;
+	output.write(`${color.gray(S_BAR_END)}  ${color.red(message)}\n\n`);
+};
+
+export const intro = (title = '', opts?: CommonOptions) => {
+	const output: Writable = opts?.output ?? process.stdout;
+	output.write(`${color.gray(S_BAR_START)}  ${title}\n`);
+};
+
+export const outro = (message = '', opts?: CommonOptions) => {
+	const output: Writable = opts?.output ?? process.stdout;
+	output.write(`${color.gray(S_BAR)}\n${color.gray(S_BAR_END)}  ${message}\n\n`);
+};
