+++ packages/prompts/src/multi-select.ts
@@ -1,0 +1,129 @@
+import { MultiSelectPrompt } from '@clack/core';
+import color from 'picocolors';
+import {
+	type CommonOptions,
+	S_BAR,
+	S_BAR_END,
+	S_CHECKBOX_ACTIVE,
+	S_CHECKBOX_INACTIVE,
+	S_CHECKBOX_SELECTED,
+	symbol,
+} from './common.js';
+import { limitOptions } from './limit-options.js';
+import type { Option } from './select.js';
+
+export interface MultiSelectOptions<Value> extends CommonOptions {
+	message: string;
+	options: Option<Value>[];
+	initialValues?: Value[];
+	maxItems?: number;
+	required?: boolean;
+	cursorAt?: Value;
+}
+export const multiselect = <Value>(opts: MultiSelectOptions<Value>) => {
+	const opt = (
+		option: Option<Value>,
+		state: 'inactive' | 'active' | 'selected' | 'active-selected' | 'submitted' | 'cancelled'
+	) => {
+		const label = option.label ?? String(option.value);
+		if (state === 'active') {
+			return `${color.cyan(S_CHECKBOX_ACTIVE)} ${label} ${
+				option.hint ? color.dim(`(${option.hint})`) : ''
+			}`;
+		}
+		if (state === 'selected') {
+			return `${color.green(S_CHECKBOX_SELECTED)} ${color.dim(label)} ${
+				option.hint ? color.dim(`(${option.hint})`) : ''
+			}`;
+		}
+		if (state === 'cancelled') {
+			return `${color.strikethrough(color.dim(label))}`;
+		}
+		if (state === 'active-selected') {
+			return `${color.green(S_CHECKBOX_SELECTED)} ${label} ${
+				option.hint ? color.dim(`(${option.hint})`) : ''
+			}`;
+		}
+		if (state === 'submitted') {
+			return `${color.dim(label)}`;
+		}
+		return `${color.dim(S_CHECKBOX_INACTIVE)} ${color.dim(label)}`;
+	};
+
+	return new MultiSelectPrompt({
+		options: opts.options,
+		input: opts.input,
+		output: opts.output,
+		initialValues: opts.initialValues,
+		required: opts.required ?? true,
+		cursorAt: opts.cursorAt,
+		validate(selected: Value[]) {
+			if (this.required && selected.length === 0)
+				return `Please select at least one option.\n${color.reset(
+					color.dim(
+						`Press ${color.gray(color.bgWhite(color.inverse(' space ')))} to select, ${color.gray(
+							color.bgWhite(color.inverse(' enter '))
+						)} to submit`
+					)
+				)}`;
+		},
+		render() {
+			const title = `${color.gray(S_BAR)}\n${symbol(this.state)}  ${opts.message}\n`;
+
+			const styleOption = (option: Option<Value>, active: boolean) => {
+				const selected = this.value.includes(option.value);
+				if (active && selected) {
+					return opt(option, 'active-selected');
+				}
+				if (selected) {
+					return opt(option, 'selected');
+				}
+				return opt(option, active ? 'active' : 'inactive');
+			};
+
+			switch (this.state) {
+				case 'submit': {
+					return `${title}${color.gray(S_BAR)}  ${
+						this.options
+							.filter(({ value }) => this.value.includes(value))
+							.map((option) => opt(option, 'submitted'))
+							.join(color.dim(', ')) || color.dim('none')
+					}`;
+				}
+				case 'cancel': {
+					const label = this.options
+						.filter(({ value }) => this.value.includes(value))
+						.map((option) => opt(option, 'cancelled'))
+						.join(color.dim(', '));
+					return `${title}${color.gray(S_BAR)}  ${
+						label.trim() ? `${label}\n${color.gray(S_BAR)}` : ''
+					}`;
+				}
+				case 'error': {
+					const footer = this.error
+						.split('\n')
+						.map((ln, i) =>
+							i === 0 ? `${color.yellow(S_BAR_END)}  ${color.yellow(ln)}` : `   ${ln}`
+						)
+						.join('\n');
+					return `${title + color.yellow(S_BAR)}  ${limitOptions({
+						output: opts.output,
+						options: this.options,
+						cursor: this.cursor,
+						maxItems: opts.maxItems,
+						style: styleOption,
+					}).join(`\n${color.yellow(S_BAR)}  `)}\n${footer}\n`;
+				}
+				default: {
+					return `${title}${color.cyan(S_BAR)}  ${limitOptions({
+						output: opts.output,
+						options: this.options,
+						cursor: this.cursor,
+						maxItems: opts.maxItems,
+						style: styleOption,
+					}).join(`\n${color.cyan(S_BAR)}  `)}\n${color.cyan(S_BAR_END)}\n`;
+				}
+			}
+		},
+	}).prompt() as Promise<Value[] | symbol>;
+};
