+++ packages/prompts/src/password.ts
@@ -1,0 +1,37 @@
+import { PasswordPrompt } from '@clack/core';
+import color from 'picocolors';
+import { type CommonOptions, S_BAR, S_BAR_END, S_PASSWORD_MASK, symbol } from './common.js';
+
+export interface PasswordOptions extends CommonOptions {
+	message: string;
+	mask?: string;
+	validate?: (value: string) => string | Error | undefined;
+}
+export const password = (opts: PasswordOptions) => {
+	return new PasswordPrompt({
+		validate: opts.validate,
+		mask: opts.mask ?? S_PASSWORD_MASK,
+		input: opts.input,
+		output: opts.output,
+		render() {
+			const title = `${color.gray(S_BAR)}\n${symbol(this.state)}  ${opts.message}\n`;
+			const value = this.valueWithCursor;
+			const masked = this.masked;
+
+			switch (this.state) {
+				case 'error':
+					return `${title.trim()}\n${color.yellow(S_BAR)}  ${masked}\n${color.yellow(
+						S_BAR_END
+					)}  ${color.yellow(this.error)}\n`;
+				case 'submit':
+					return `${title}${color.gray(S_BAR)}  ${color.dim(masked)}`;
+				case 'cancel':
+					return `${title}${color.gray(S_BAR)}  ${color.strikethrough(color.dim(masked ?? ''))}${
+						masked ? `\n${color.gray(S_BAR)}` : ''
+					}`;
+				default:
+					return `${title}${color.cyan(S_BAR)}  ${value}\n${color.cyan(S_BAR_END)}\n`;
+			}
+		},
+	}).prompt() as Promise<string | symbol>;
+};
